* commit '410a1628d13845ac7ffeb69f9598e5b9f6b9906f':
  Improve the spot shadow computation.

+++ libs/hwui/Interpolator.cpp
@@ -110,7 +110,7 @@ float LUTInterpolator::interpolate(float input) {
     weight = modff(lutpos, &ipart);
 
     int i1 = (int) ipart;
-    int i2 = MathUtils::min(i1 + 1, mSize - 1);
+    int i2 = MathUtils::min(i1 + 1, (int) mSize - 1);
 
     LOG_ALWAYS_FATAL_IF(i1 < 0 || i2 < 0, "negatives in interpolation!"
             " i1=%d, i2=%d, input=%f, lutpos=%f, size=%zu, values=%p, ipart=%f, weight=%f",
